Finish chapter 14
Note: there seems to be a tiny bug in the server: when you
do ./server -i, it initializes the dbs, but fails to exit. I had
to hit ctl-c.

Not sure why this happens with the ch14 version but not the ch13 one...
I thought the logic path for command line didn't even touch the
code that changed.